Cracked meshless components into its own library, separate from model.
