+++ meshless/interpreter_register_meshless_functions.c
@@ -1,0 +1,18 @@
+// Copyright (c) 2012-2015, Jeffrey N. Johnson
+// All rights reserved.
+// 
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+#include "meshless/interpreter_register_meshless_functions.h"
+
+// Lua stuff.
+#include "lua.h"
+#include "lualib.h"
+#include "lauxlib.h"
+
+// This is currently just a placeholder.
+void interpreter_register_meshless_functions(interpreter_t* interp)
+{
+}
