+++ model/tests/CMakeLists.txt
@@ -3,11 +3,6 @@ include(add_polymec_test)
 add_polymec_test(test_polynomial_fit test_polynomial_fit.c)
 add_polymec_test(test_neighbor_pairing test_neighbor_pairing.c create_simple_pairing.c)
 add_mpi_polymec_test(test_stencil test_stencil.c 1 2 3 4)
-add_mpi_polymec_test(test_partition_point_cloud_with_neighbors test_partition_point_cloud_with_neighbors.c create_simple_pairing.c 1 2 3 4)
-add_mpi_polymec_test(test_shepard_shape_function test_shepard_shape_function.c 1 2 3 4)
-add_mpi_polymec_test(test_mls_shape_function test_mls_shape_function.c 1 2 3 4)
-add_polymec_test(test_gmls_functional test_gmls_functional.c poisson_gmls_functional.c make_mlpg_lattice.c)
-add_polymec_test(test_gmls_matrix test_gmls_matrix.c poisson_gmls_functional.c make_mlpg_lattice.c)
 
 function(add_game_of_life_test input num_steps diff)
   if (HAVE_MPI EQUAL 1)
